requirements.txt
Flask==2.0.1
numpy==1.21.0
pillow==8.2.0
torch==1.9.0
torchvision==0.10.0
